requirements.txt
jax
jaxlib
flax
optax
tensorflow-datasets
tensorflow-macos
tensorflow-metal